requirements.txt
python-telegram-bot==21.9
openai==1.65.0
asyncpg==0.30.0
python-dotenv==1.0.1
httpx==0.27.2
